feat: apply change needed to be deleted
